fix(canary): racey canary service account deletion
This occured when a kafka that has not been prepared is deleted. The root cause is that for these
kafkas, the canary service account id is empty, when retrieving the the internalClientId, the
keycloak client returned the first id in the list of all ids it retreved because the filtering which
is based on the Kafka.CanaryServiceAccountClientId was an empty filters due to the id being empty.

+++ internal/kafka/internal/services/kafka.go
@@ -459,9 +459,12 @@ func (k *kafkaService) Delete(kafkaRequest *dbapi.KafkaRequest) *errors.ServiceE
 			if keycloakErr != nil {
 				return errors.NewWithCause(errors.ErrorGeneral, keycloakErr, "error deleting sso client")
 			}
-			keycloakErr = k.keycloakService.DeleteServiceAccountInternal(kafkaRequest.CanaryServiceAccountClientID)
-			if keycloakErr != nil {
-				return errors.NewWithCause(errors.ErrorGeneral, keycloakErr, "error deleting canary service account")
+
+			if kafkaRequest.CanaryServiceAccountClientID != "" {
+				keycloakErr = k.keycloakService.DeleteServiceAccountInternal(kafkaRequest.CanaryServiceAccountClientID)
+				if keycloakErr != nil {
+					return errors.NewWithCause(errors.ErrorGeneral, keycloakErr, "error deleting canary service account")
+				}
 			}
 		}
